+++ checkpoint/orbax/__init__.py
@@ -19,7 +19,6 @@ import functools
 
 from orbax.checkpoint import aggregate_handlers
 from orbax.checkpoint import checkpoint_utils
-from orbax.checkpoint import lazy_utils
 from orbax.checkpoint import msgpack_utils
 from orbax.checkpoint import test_utils
 from orbax.checkpoint import transform_utils