+++ checkpoint/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Allow calling `create_coordinator_server_and_context` without an initialized
 JAX coordinator server and create TS metadata for numpy arrays on a single process.
 - Refactor TypeHandler to operate over batches of values, rather than individual ones.
+- Removed support for lazy restoration. Supported via transformations.
 
 ## [0.2.6] - 2022-06-28